// src/app/shared/models/employee.ts
export interface EmployeeReturn
 {
    id: number;
    name: string;
    address: string;
    birthDate: Date;
    weChat: string;
    hireDate: Date;
    sickLeave: number;
    payRate: number;
    annualLeave: number;
    superAccount: string;
    departmentName: string;
    departmentId:number;
}